Feat/route location

+++ migrations/20171124153521_bike_app.js
@@ -42,7 +42,7 @@ exports.up = function(knex, Promise) {
         .references('id')
         .inTable('route');
       t.decimal('mph');
-      t.integer('time');
+      t.decimal('time');
       t.decimal('distance');
       t.integer('calories');
     }),
